Creates the EntityFramework package.
