+++ PosInformatique.UnitTests.Databases.sln
@@ -44,6 +44,10 @@ Project("{00D1A9C2-B5F0-4AF3-8072-F6C62B433612}") = "UnitTests.Databases.SqlServ
 EndProject
 Project("{00D1A9C2-B5F0-4AF3-8072-F6C62B433612}") = "UnitTests.Databases.SqlServer.Tests.Target", "tests\UnitTests.Databases.SqlServer.Tests.Target\UnitTests.Databases.SqlServer.Tests.Target.sqlproj", "{6CD3F177-053F-4816-A37E-5CA6F293D34C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests.Databases.SqlServer.EntityFramework", "src\UnitTests.Databases.SqlServer.EntityFramework\UnitTests.Databases.SqlServer.EntityFramework.csproj", "{157DDF0D-9410-4646-94B9-9CEE4C140F5E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTests.Databases.SqlServer.EntityFramework.Tests", "tests\UnitTests.Databases.SqlServer.EntityFramework.Tests\UnitTests.Databases.SqlServer.EntityFramework.Tests.csproj", "{04A7AE8F-FE77-435B-9250-600388BB8065}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -76,6 +80,14 @@ Global
 		{6CD3F177-053F-4816-A37E-5CA6F293D34C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6CD3F177-053F-4816-A37E-5CA6F293D34C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6CD3F177-053F-4816-A37E-5CA6F293D34C}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{157DDF0D-9410-4646-94B9-9CEE4C140F5E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{157DDF0D-9410-4646-94B9-9CEE4C140F5E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{157DDF0D-9410-4646-94B9-9CEE4C140F5E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{157DDF0D-9410-4646-94B9-9CEE4C140F5E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{04A7AE8F-FE77-435B-9250-600388BB8065}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{04A7AE8F-FE77-435B-9250-600388BB8065}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{04A7AE8F-FE77-435B-9250-600388BB8065}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{04A7AE8F-FE77-435B-9250-600388BB8065}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
