+++ src/UnitTests.Databases.SqlServer/UnitTests.Databases.SqlServer.csproj
@@ -5,7 +5,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.0" />
     <PackageReference Include="Microsoft.SqlServer.DacFx" Version="162.3.566" />
   </ItemGroup>
